update: atualiza arquivos e prepara análise de tempo para o app
