+++ pyproject.toml
@@ -20,6 +20,11 @@ mkdocs-video = "^1.3.0"
 mkdocs-glightbox = "^0.1.7"
 pre-commit = "^2.20.0"
 seaborn = "^0.11.2"
+streamlit-tags = "^1.2.8"
+snscrape = "^0.4.3"
+nltk = "^3.7"
+plotly = "^5.11.0"
+bokeh = "2.4.3"
 
 [tool.poetry.dev-dependencies]
 pytest = "^5.2"
